+++ CHANGES.md
@@ -1,5 +1,16 @@
 # Calendar.js - Change Log:
 
+### Version 0.9.3:
+- Added export to MD (markdown) file support.
+- Added export to HTML file support.
+- Added public function getVersion(), which returns the current version of Calendar.js.
+- Fixed some documentation errors for functions (return types not being shown correctly).
+- Fixed the PRODID for iCAL exports showing the wrong product data.
+- Added gregorian time format support for iCAL export type.
+- Added new option tooltipsEnabled, which allows the tooltips throughout the calendar to be disabled (on by default).
+- Fixed a fault that caused the holidays to appear within the Day number text element (now placed under it).
+- The first day of each month is now highlighted (the same style as the current day of the month).
+
 ### Version 0.9.2:
 - Added support to show the previous/next month names in the days on the main display (optional, on by default).
 - Added a new option "showDayNamesInMainDisplay" to disable the day names header in the main display (optional, on by default).